Fix most of the color bleeding by capping values in the dithering accumulation buffer above instead of just below
Also, cap values only the end instead of after each accumulation; both faster
and fixes even more artifacts.
But allow values to go a bit below 0 and above 255 (controlled by kAccumMargin,
which I set to 64 but I didn't experiment with different values), so as to not
throw away large errors. This seems to improve results.