Fix erroneous k-d tree implementation
The way it was implemented (before and after jacobly0's changes), points lying
on a splitting hyperplane would be randomly put in either child. But the whole
point of a k-d tree is that you know which child needs to be checked. Aside from
being less efficient, it also caused unnecessary mixing of colours. It could
select a palette of X colours from an image with Y colours, with Y << X.

So, this provides another speedup to finding the nearest colour.