Replace inefficient recursive k-d tree splitting algorithm with the standard algorithm for building a palette
The difference is that instead of ensuring that each palette colour corresponds
to equal numbers of pixels with a balanced k-d tree, which is awful if most of
the image is a single color, the node of the k-d tree which has the largest
cost (quantisation error) is split. This uses a priority queue instead of
the simpler recursive splitting.

The cost metric could probably be improved substantially (variance instead
of max range?)